[Refactor] import 경로수정, console.log 제거
