+++ src/app/(default)/donate/[id]/layout.tsx
@@ -1,4 +1,5 @@
-import DetailInfo, { DetailInfoProps } from "@/shared/components/DetailInfo";
+import ParticipationDetailInfo from "@/domain/participation/components/ParticipationDetailInfo";
+import { DetailInfoProps } from "@/domain/participation/types";
 import ImageSwiper, { dummyImageList } from "@/shared/components/ImageSwiper";
 import TabBar from "@/shared/components/TabBar";
 import { donateTabContents } from "@/shared/utils/navigation";
@@ -33,7 +34,7 @@ async function page({
         <TabBar type="donate" tabContents={donateTabContents(id)} />
         <main className="py-4">{children}</main>
       </div>
-      <DetailInfo props={DETAIL_INFO_DUMMY} />
+      <ParticipationDetailInfo props={DETAIL_INFO_DUMMY} />
     </div>
   );
 }
