+++ src/domain/participation/components/ProgressBar.tsx
@@ -1,4 +1,4 @@
-import { formatMoney } from "@/shared/utils/formatMoney";
+import { formatMoney } from "@/domain/donate/utils/formatMoney";
 
 interface ProgressBarProps {
   type: "donate" | "together";
