+++ src/domain/participation/utils/syncUrl.ts
